refactor(api): isolate run idempotency storage
